goap doesnt work anymore

+++ BehaviourTreeExample/Assets/Scripts/Interfaces/IDamageable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(GameObject attacker, int damage);
+    void TakeDamage(GameObject attacker, float damage);
 }
